Complete Map
Last push of work for the map, including a basic landing zone (though it currently only goes landing -> map). Also pushing a few files I was fiddling with to try to add in custom hair; it didn't work but it didn't hurt either so I'm including them in the commit.